[Add][Doc] Gaggle example
What?
A self contained example of a gaggle setup.

Why?
This assists end to end testing.
It also suggests Goose functionality that is required for
end-to-end testing a gaggle configuration.
Specifically the requirement to provide IP-address with
`--master-host ...` blocks such test infrastructure.

How?
Provides a Docker or Podman or k3s container cluster.

Signed-off-by: Begley Brothers Ltd <begleybrothers@gmail.com>